Properties are done (only lack of internal code beautification for the PropertiesTrait.php), and test coverage was adjusted back to 100%.
SystemFingerprints seems to be working, but are not fully done yet.

Though Test coverage is 100%, there are some sketchy places that has to be improved at some point, and they might have been ignored for coverage (not a lot of such).

Documentation must be fully revised! :(

Further code improvements are required:
 * Tests
 * PropertiesTrait.php
 * Overall php fields related to Traits must be moved into those traits
 * General revision and adjustment of the architecture

6163e493c070382251361366 Finalizing new-features-2 PR/MR